requirements.txt
Pillow==10.3.0
Requests==2.31.0
streamlit==1.31.1

